requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.4.1
astroid==2.8.0
backports.entry-points-selectable==1.1.0
colorama==0.4.4
cloudinary==1.26.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-filter==2.4.0
django-userforeignkey==0.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
imutils==0.5.4
isort==5.9.3
join==0.1.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.3.2
platformdirs==2.3.0
psycopg2==2.9.1
pylint==2.11.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.4
pytz==2021.1

six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
virtualenv==20.8.0
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.12.1
xhtml2pdf==0.2.5

django-cloudinary-storage==0.3.0